requirements.txt
# ============================================================================
# VAT_Audit_Project 依赖包清单
# 更新日期: 2026-01-04
# Python 版本要求: >= 3.8
# ============================================================================

# 核心数据处理库
# ----------------------------------------------------------------------------
pandas>=2.0.0          # DataFrame 操作、数据清洗、CSV/Excel 读写
numpy>=1.24.0          # 数值计算、数组操作（pandas 依赖）

# Excel 文件处理
# ----------------------------------------------------------------------------
openpyxl>=3.0.0        # 读写 .xlsx 文件（Excel 2010+）
xlrd==2.0.1            # 读取 .xls 文件（旧版 Excel）
                       # 注意: pandas >=2 需要 2.0.1+ 才能读取 .xls

# 编码检测与处理
# ----------------------------------------------------------------------------
chardet>=5.0.0         # 自动检测文件编码（支持 GB2312/UTF-8/GBK 等）

# 系统监控与性能优化
# ----------------------------------------------------------------------------
psutil>=5.9.0          # CPU/内存监控、进程管理、系统信息获取
tqdm>=4.65.0           # 进度条显示（用于文件处理进度提示）

# 命令行接口
# ----------------------------------------------------------------------------
click>=8.1.0           # CLI（run/analyze 命令）

# 配置文件解析
# ----------------------------------------------------------------------------
pyyaml>=6.0.1         # YAML 配置解析

# 打包工具（仅用于生成 .exe）
# ----------------------------------------------------------------------------
# pyinstaller>=6.0.0    # 打包成可执行文件（开发时安装）
                        # 安装命令: pip install pyinstaller

# ============================================================================
# 内置库（无需安装）
# ============================================================================
# 以下库为 Python 标准库，无需额外安装：
# - sqlite3        : SQLite 数据库支持
# - logging        : 日志系统
# - multiprocessing: 多进程并行处理
# - os, sys, re    : 系统和文件操作
# - datetime       : 日期时间处理
# - json           : JSON 数据处理
# - uuid           : 唯一标识符生成
# - threading      : 线程支持
# - tempfile       : 临时文件管理
# - shutil         : 文件操作工具
# - glob           : 文件路径匹配
# - atexit         : 程序退出处理
# - dataclasses    : 数据类支持
# - typing         : 类型注解支持
# - collections    : 集合类型扩展
# - contextlib     : 上下文管理器工具
# - enum           : 枚举类型支持
